# src/hooks/use-custom-api.ts
import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
import { useMemo } from "react";

// Base URL for our custom API
const API_BASE = import.meta.env.VITE_API_URL || "http://localhost:3001/api";

// Types (will be generated by orval, but defining here for now)
export interface CustomPlanet {
  id: number;
  planetId: string;
  customNotes?: string;
  userRating?: number;
  isFavorite: boolean;
  customImage?: string;
  tags?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CustomFilm {
  id: number;
  filmId: string;
  customNotes?: string;
  userRating?: number;
  isFavorite: boolean;
  watchedCount: number;
  createdAt: string;
  updatedAt: string;
}

// Custom Planet Hooks
export function useCustomPlanet(planetId: string | null) {
  return useQuery<CustomPlanet | null>({
    queryKey: ["custom-planet", planetId],
    queryFn: async () => {
      if (!planetId) return null;
      console.time(`[CustomAPI] getPlanet ${planetId}`);
      const res = await fetch(`${API_BASE}/custom/planets?planetId=${planetId}`);
      if (!res.ok) throw new Error("Failed to fetch custom planet data");
      const data = await res.json();
      console.timeEnd(`[CustomAPI] getPlanet ${planetId}`);
      return data;
    },
    enabled: !!planetId,
    staleTime: 1000 * 60 * 5, // 5 minutes
  });
}

export function useCustomPlanets() {
  return useQuery<CustomPlanet[]>({
    queryKey: ["custom-planets"],
    queryFn: async () => {
      console.time("[CustomAPI] getAllPlanets");
      const res = await fetch(`${API_BASE}/custom/planets`);
      if (!res.ok) throw new Error("Failed to fetch custom planets");
      const data = await res.json();
      console.timeEnd("[CustomAPI] getAllPlanets");
      return data;
    },
    staleTime: 1000 * 60 * 5,
  });
}

export function useUpsertCustomPlanet() {
  const queryClient = useQueryClient();
  
  return useMutation({
    mutationFn: async (data: Partial<CustomPlanet> & { planetId: string }) => {
      console.time("[CustomAPI] upsertPlanet");
      const res = await fetch(`${API_BASE}/custom/planets`, {
        method: "POST",
        headers: { "Content-Type": "application/json" },
        body: JSON.stringify(data),
      });
      if (!res.ok) throw new Error("Failed to save custom planet data");
      const result = await res.json();
      console.timeEnd("[CustomAPI] upsertPlanet");
      return result;
    },
    onSuccess: (data) => {
      queryClient.invalidateQueries({ queryKey: ["custom-planet", data.planetId] });
      queryClient.invalidateQueries({ queryKey: ["custom-planets"] });
      console.log("[CustomAPI] Planet saved:", data);
    },
    onError: (error) => {
      console.error("[CustomAPI] Error saving planet:", error);
    },
  });
}

// Custom Film Hooks
export function useCustomFilm(filmId: string | null) {
  return useQuery<CustomFilm | null>({
    queryKey: ["custom-film", filmId],
    queryFn: async () => {
      if (!filmId) return null;
      console.time(`[CustomAPI] getFilm ${filmId}`);
      const res = await fetch(`${API_BASE}/custom/films?filmId=${filmId}`);
      if (!res.ok) throw new Error("Failed to fetch custom film data");
      const data = await res.json();
      console.timeEnd(`[CustomAPI] getFilm ${filmId}`);
      return data;
    },
    enabled: !!filmId,
    staleTime: 1000 * 60 * 5,
  });
}

export function useUpsertCustomFilm() {
  const queryClient = useQueryClient();
  
  return useMutation({
    mutationFn: async (data: Partial<CustomFilm> & { filmId: string }) => {
      console.time("[CustomAPI] upsertFilm");
      const res = await fetch(`${API_BASE}/custom/films`, {
        method: "POST",
        headers: { "Content-Type": "application/json" },
        body: JSON.stringify(data),
      });
      if (!res.ok) throw new Error("Failed to save custom film data");
      const result = await res.json();
      console.timeEnd("[CustomAPI] upsertFilm");
      return result;
    },
    onSuccess: (data) => {
      queryClient.invalidateQueries({ queryKey: ["custom-film", data.filmId] });
      queryClient.invalidateQueries({ queryKey: ["custom-films"] });
      console.log("[CustomAPI] Film saved:", data);
    },
    onError: (error) => {
      console.error("[CustomAPI] Error saving film:", error);
    },
  });
}

// Helper to merge custom data with SWAPI data
export function useMergedPlanetWithCustom(
  planet: { id: string; name: string } | undefined,
) {
  const { data: customData } = useCustomPlanet(planet?.id || null);
  
  return useMemo(() => {
    if (!planet) return undefined;
    
    return {
      ...planet,
      customNotes: customData?.customNotes,
      userRating: customData?.userRating,
      isFavorite: customData?.isFavorite || false,
      customImage: customData?.customImage,
      tags: customData?.tags,
    };
  }, [planet, customData]);
}